feat: Divide InformationCard to Tour & Rest Card
